Extract interpreter to different module
